move fake type out to own file
